+++ src/XrmContext/Interpretation/InterpretIntersectionInterfaces.fs
@@ -1,0 +1,12 @@
+﻿namespace DG.XrmContext
+
+open System
+open IntermediateRepresentation
+open InterpretOptionSetMetadata
+open Microsoft.Xrm.Sdk
+open Microsoft.Xrm.Sdk.Metadata
+open Utility
+
+module internal InterpretIntersectionInterfaces =
+  
+  let interpretIntersectInterfaces = ()