+++ src/XrmContext/XrmContext.fs
@@ -2,65 +2,111 @@
 
 open System
 open Utility
-open GeneratorLogic
+open FileGeneration
+open GenerationMain
+
 open System.Diagnostics
 open System.IO
+open System.Runtime.Serialization.Json
 
 type XrmContext private () =
 
-  static member GetContext(url, username, password, ?domain, ?ap, ?out, ?entities, ?solutions, ?ns, ?context, ?deprecatedPrefix, ?sdkVersion) =
-    let xrmAuth =
+  static member GenerateFromCrm(url, username, password, ?domain, ?ap, ?out, ?entities, ?solutions, ?ns, ?context, ?deprecatedPrefix, ?sdkVersion, ?intersections) = 
+    let xrmAuth = 
       { XrmAuthentication.url = Uri(url)
         username = username
         password = password
         domain = domain
-        ap = ap
+        ap = ap }
+    
+    let rSettings = 
+      { XcRetrievalSettings.entities = entities
+        solutions = solutions
       }
 
-    let settings = 
-      { XrmContextSettings.out = out
+    let gSettings = 
+      { XcGenerationSettings.out = out
+        sdkVersion = sdkVersion
+        deprecatedPrefix = deprecatedPrefix
         ns = ns
         context = context
-        solutions = solutions
-        entities = entities
-        deprecatedPrefix = deprecatedPrefix
-        sdkVersion = sdkVersion
-      }
+        intersections = intersections
+       }
     
-    XrmContext.GetContext(xrmAuth, settings)
+    XrmContext.GenerateFromCrm(xrmAuth, rSettings, gSettings)
+  
 
 
-  static member GetContext(xrmAuth, settings) =
-    #if !DEBUG
+  static member GenerateFromCrm(xrmAuth, rSettings, gSettings) =
+    #if !DEBUG 
     try
-    #endif
-      let out = settings.out |? "."
-      let ns = settings.ns |? null
-
-      // Connect to CRM and interpret the data
-      let proxy = connectToCrm xrmAuth
+    #endif 
       
-      let proxyGetter = proxyHelper xrmAuth
-      let entities = 
-        getFullEntityList settings.entities settings.solutions proxy
-      
-      let sdkVersion =
-        if settings.sdkVersion.IsNone then retrieveCrmVersion proxy
-        else settings.sdkVersion.Value
-
-      let data = 
-        (proxy, proxyGetter)
-        ||> retrieveCrmData entities
-        |> interpretCrmData out ns settings.deprecatedPrefix settings.context sdkVersion
-
-      // Generate the code
-      createCodeDom data
-      createResourceFiles sdkVersion out
-
-      printfn "\nDone generating context!"
+      retrieveRawState xrmAuth rSettings
+      |> generateFromRaw gSettings
+      printfn "\nSuccessfully generated the C# context files."
 
     #if !DEBUG
-    with _ as ex ->
-      getFirstExceptionMessage ex |> failwithf "\nUnable to generate context: %s"
+    with ex -> getFirstExceptionMessage ex |> failwithf "\nUnable to generate context files: %s"
     #endif
 
+
+
+  static member SaveMetadataToFile(xrmAuth, rSettings, ?filePath) =
+    #if !DEBUG 
+    try
+    #endif 
+      
+      let filePath = 
+        filePath 
+        ?>>? (String.IsNullOrWhiteSpace >> not)
+        ?| "XcData.json"
+
+      let serializer = DataContractJsonSerializer(typeof<RawState>)
+      use stream = new FileStream(filePath, FileMode.Create)
+
+      retrieveRawState xrmAuth rSettings
+      |> fun state -> serializer.WriteObject(stream, state)
+      printfn "\nSuccessfully saved retrieved data to file."
+
+    #if !DEBUG
+    with ex -> getFirstExceptionMessage ex |> failwithf "\nUnable to generate context file: %s"
+    #endif
+
+
+
+  static member GenerateFromRawState(rawState, gSettings) =
+    #if !DEBUG 
+    try
+    #endif 
+
+      generateFromRaw gSettings rawState
+      printfn "\nSuccessfully generated the C# context files."
+
+    #if !DEBUG
+    with ex -> getFirstExceptionMessage ex |> failwithf "\nUnable to generate context files: %s"
+    #endif
+
+
+  static member GenerateFromFile(gSettings, ?filePath) =
+    #if !DEBUG 
+    try
+    #endif 
+      let filePath = 
+        filePath 
+        ?>>? (String.IsNullOrWhiteSpace >> not)
+        ?| "XcData.json"
+
+      let rawState =
+        try
+          let serializer = DataContractJsonSerializer(typeof<RawState>)
+          use stream = new FileStream(filePath, FileMode.Open)
+          serializer.ReadObject(stream) :?> RawState
+        with ex -> failwithf "\nUnable to parse data file"
+
+      generateFromRaw gSettings rawState
+      printfn "\nSuccessfully generated the C# context files."
+
+    #if !DEBUG
+    with ex -> getFirstExceptionMessage ex |> failwithf "\nUnable to generate context files: %s"
+    #endif